+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - [#2756](https://github.com/poanetwork/blockscout/pull/2756) - Improve subquery joins
 
 ### Chore
+- [#3099](https://github.com/poanetwork/blockscout/pull/3099) - Remove pending txs cache
 - [#3093](https://github.com/poanetwork/blockscout/pull/3093) - Extend list of env vars for Docker setup
 - [#3084](https://github.com/poanetwork/blockscout/pull/3084) - Bump Elixir version 1.10.2
 - [#3079](https://github.com/poanetwork/blockscout/pull/3079) - Extend optionality of websockets to Geth